feat(account): migrate proto to common/rust/shared-proto, delete username and pass integration tests
* feat(account): migrate proto to common/rust/shared-proto, delete username and pass integration tests

* feat(account): add identity handler

* refactor(account): remove regions in usescases

* refactor(account): remove regions in usescases, refactor tests and implement fixure (compile not working)

* refactor(account): refactor and fix usecases units test, impl context for tests

* refactor: standardize aggregate root pattern and lifecycle

Move logic from entities to Account aggregate root.

Implement scoped AccountContext for transactional safety.

Align gRPC proto with domain value objects (oneof identifier).

Update test fixtures and repository stubs for aggregate-based testing.

* test_worst_case_concurrency_conflict

* refactor(account): enhance repository resilience and performance

Add Redis caching with non-blocking fail-safe (spawn & timeout)

Implement Optimistic Concurrency Control (OCC) to prevent data race

Fix transaction atomicity and cache invalidation logic

Add comprehensive integration tests for lifecycle and concurrency

---------

+++ crates/account/src/application/use_cases/moderation/decrease_trust_score/decrease_trust_score_use_case.rs
@@ -18,7 +18,7 @@ impl CommandHandler for DecreaseTrustScoreHandler {
 
     async fn handle(&self, ctx: &AccountContext, cmd: DecreaseTrustScoreCommand) -> Result<Self::Output> {
         let mut account = ctx.account().await?;
-        account.penalize_trust(cmd.amount as i32, cmd.reason)?;
+        account.penalize_trust(cmd.amount, cmd.reason)?;
         ctx.save(&mut account, Some(cmd.command_id)).await?;
         Ok(())
     }
